fix sprites not being rendered on lowest possible floor
Applied: https://github.com/fabiangreffrath/crispy-doom/commit/6d81ad8567d8576ef9c83dae035488588bffbb18
Thanks @fabiangreffrath!

+++ src/doom/r_things.c
@@ -652,7 +652,7 @@ void R_ProjectSprite (mobj_t* thing)
     gzt = interpz + spritetopoffset[lump];
 
     if (interpz > viewz + FixedDiv(viewheight << FRACBITS, xscale)
-    ||  gzt     < viewz - FixedDiv((viewheight << FRACBITS)-viewheight, xscale))
+    ||  gzt     < (int64_t)viewz - FixedDiv((viewheight << FRACBITS)-viewheight, xscale))
     return;
 
     // [JN] quickly reject sprites with bad x ranges
